add mock functions for adding and removing todo tasks
